Move the Builder into a new module CLI
This keeps the CLI builder tools separate from the Runner methods
