+++ lib/commander.rb
@@ -31,6 +31,5 @@ require 'commander/runner'
 require 'commander/command'
 require 'commander/help_formatters'
 require 'commander/platform'
-require 'commander/delegates'
-require 'commander/methods'
+require 'commander/cli'
 require 'commander/configure'
