--- lib/commander/delegates.rb
@@ -1,4 +1,0 @@
-module Commander
-  # TODO: Replace this fully with the Builder
-  Delegates = Commander::Runner::Builder
-end
